requirements.txt
streamlit>=1.24.0
langchain>=0.0.267
langchain-community>=0.0.10
python-dotenv>=1.0.0
openai>=1.3.0
praw>=7.7.1
snscrape>=0.7.0.20221206
chromadb>=0.4.18
watchdog>=3.0.0
pydantic>=2.0.0
wordcloud>=1.9.2 